refactor: Update namespaces and import statements for AddToCartWithOptions blocks
