+++ plugins/woocommerce/src/Blocks/BlockTypes/AddToCartWithOptions/AddToCartWithOptionsGroupedProductSelectorItemTemplate.php
@@ -1,8 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace Automattic\WooCommerce\Blocks\BlockTypes;
+namespace Automattic\WooCommerce\Blocks\BlockTypes\AddToCartWithOptions;
 
+use Automattic\WooCommerce\Blocks\BlockTypes\AbstractBlock;
+use Automattic\WooCommerce\Blocks\BlockTypes\EnableBlockJsonAssetsTrait;
 use WP_Block;
 
 /**
